Extract modules
Extract XxModule classes from ArrayListModule class.
