+++ BasketballModule.java
@@ -1,0 +1,15 @@
+package JavaReport;
+
+import java.util.ArrayList;
+
+public class BasketballModule {
+    ArrayList<String> ArrayListBasketball = new ArrayList<String>();
+
+    public BasketballModule() {
+        ArrayListBasketball.add("(마이클 조던) : 두려움은 환상이다 \n\n");
+        ArrayListBasketball.add("9000번 이상 슛을 놓치고.300번의 게임에서 졌다.\n 나를 믿고 맡겨진 결정골을 놓친것은 26번이다.\n 나는 지금까지 실패하고 또 실패했다.\n 그것이 내가 성공한 이유다.\n\n");
+        ArrayListBasketball.add("WELCOME to NBA \n\n\n");
+        ArrayListBasketball.add("(찰스 바클리) : 누군가는 영웅이 되어야 한다. 어쩌면 그 영웅이 나일지도 모른다 \n\n");
+        System.out.println(ArrayListBasketball);
+    }
+}
