+++ DeveloperModule.java
@@ -1,0 +1,18 @@
+package JavaReport;
+
+import java.util.ArrayList;
+
+public class DeveloperModule {
+    ArrayList<String> ArrayListProgramming = new ArrayList<String>();
+
+    public DeveloperModule() {
+        ArrayListProgramming.add("(코딩의 신 아샬) : 안녕하세요, 코딩의 신 아샬입니다. \n\n");
+        ArrayListProgramming.add("안될 때마다 속으로 외치세요. \n뭔가 잘못했겠지. 뭔가 잘못했겠지. \n\n");
+        ArrayListProgramming.add("갑자기 산에 틀어 박혀서 \n난 프로그래밍의 극의에 다다른 자 으아! \n이러고 내려와서 실전에 가보면\n발릴 걸요? \n그냥 발릴 걸요?\n쪽도 못 쓸 걸요?\n이건 중 2병이죠.\n\n");
+        ArrayListProgramming.add("피아노는 이렇게 손바닥을 펴고 때리는 거야! 이렇게! 도레미파솔!    \n\n");
+        ArrayListProgramming.add("Managed 언어, Unmanaged 언어 이런 게 중요한 게 아니라\n컴퓨터를 알아야 돼요 컴퓨터를!  \n\n");
+        ArrayListProgramming.add("으아아 개XX들아! 난 나간다 으아아! \n\n");
+        ArrayListProgramming.add("그럼 다들 안녕히 주무세요. 바이~ \n\n");
+        System.out.println(ArrayListProgramming);
+    }
+}
