+++ ViewBasketball.java
@@ -7,9 +7,7 @@ public class ViewBasketball {
         Scanner chooseSports = new Scanner(System.in);
         int selectedSports = chooseSports.nextInt();
         if (selectedSports == 1) {
-            // Reference by ArrayListModule.java
-            ArrayListModule LargeModule = new ArrayListModule();
-            ArrayListModule.BasketballModule showBasketball = LargeModule.new BasketballModule();
+            BasketballModule showBasketball = new BasketballModule();
             System.out.println(showBasketball);
         }
     }
