+++ ViewDeveloper.java
@@ -7,9 +7,7 @@ public class ViewDeveloper {
         Scanner chooseProgramming = new Scanner(System.in);
         int selectedProgramming = chooseProgramming.nextInt();
         if (selectedProgramming == 1) {
-            // Reference by ArrayListModule.java
-            ArrayListModule LargeModule = new ArrayListModule();
-            ArrayListModule.DeveloperModule showDeveloper = LargeModule.new DeveloperModule();
+            DeveloperModule showDeveloper = new DeveloperModule();
             System.out.println(showDeveloper);
         }
     }
