+++ ViewEntrepreneur.java
@@ -7,9 +7,7 @@ public class ViewEntrepreneur {
         Scanner chooseBusiness = new Scanner(System.in);
         int selectedBusiness = chooseBusiness.nextInt();
         if (selectedBusiness == 1) {
-            // Reference by ArrayListModule.java
-            ArrayListModule LargeModule = new ArrayListModule();
-            ArrayListModule.EntrepreneurModule showEntrepeneur = LargeModule.new EntrepreneurModule();
+            EntrepreneurModule showEntrepeneur = new EntrepreneurModule();
             System.out.println(showEntrepeneur);
         }
     }
